+++ packages/blade/src/components/Typography/BaseText/getBaseTextStyles.ts
@@ -17,7 +17,7 @@ const getBaseTextStyles = ({
   font-style: ${fontStyle};
   text-decoration-line: ${textDecorationLine};
   line-height: ${lineHeight};
-  ${textAlign ? `text-align: ${textAlign};` : ''}
+  ${textAlign ? `text-align: ${textAlign};` : ''};
 `;
 
 export default getBaseTextStyles;