+++ packages/blade/src/components/VisuallyHidden/VisuallyHidden.native.tsx
@@ -5,8 +5,6 @@ import type { VisuallyHiddenProps } from './VisuallyHidden.d';
 
 const StyledVisuallyHidden = styled.View(screenReaderStyles);
 
-const VisuallyHidden = ({ children }: VisuallyHiddenProps): JSX.Element => {
+export const VisuallyHidden = ({ children }: VisuallyHiddenProps): JSX.Element => {
   return <StyledVisuallyHidden>{children}</StyledVisuallyHidden>;
 };
-
-export default VisuallyHidden;