+++ packages/blade/src/utils/makeSpace/makeSpace.test.ts
@@ -1,0 +1,9 @@
+import spacing from '../../tokens/global/spacing';
+import makeSpace from '.';
+
+describe('makeSpace', () => {
+  it('should return the spacing value in `px`', () => {
+    const space = makeSpace(spacing[2]);
+    expect(space).toEqual('8px');
+  });
+});